fix: rename sport-event in app module
